+++ smol-bytes/src/bytes/raw/serde.rs
@@ -32,7 +32,7 @@ where
       A: serde_core::de::SeqAccess<'a>,
     {
       let mut values = match seq.size_hint() {
-        Some(hint) => BytesMut::with_capacity(hint),
+        Some(hint) => BytesMut::with_capacity(hint.min(4096)),
         None => BytesMut::new(),
       };